Support for 64k globals in progs.dat.

+++ source/pr_exec.c
@@ -602,16 +602,16 @@ while (1)
 
 	case OP_IFNOT:
 		if (!a->_int)
-			s += st->b - 1;	// offset the s++
+			s += (signed short)st->b - 1;	// offset the s++
 		break;
 		
 	case OP_IF:
 		if (a->_int)
-			s += st->b - 1;	// offset the s++
+			s += (signed short)st->b - 1;	// offset the s++
 		break;
 		
 	case OP_GOTO:
-		s += st->a - 1;	// offset the s++
+		s += (signed short)st->a - 1;	// offset the s++
 		break;
 		
 	case OP_CALL0:
